# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: ThePrivateLedger API
  version: 0.2.0
  description: API for managing financial accounts, transactions, and performance
    metrics.
  contact:
    name: Bellanov Apilli
    email: bellanovapilli@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /accounts:
    get:
      summary: Get Accounts
      operationId: get_accounts_accounts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Account'
                type: array
                title: Response Get Accounts Accounts Get
  /accounts/{account_id}:
    get:
      summary: Get Account
      operationId: get_account_accounts__account_id__get
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
                title: Response Get Account Accounts  Account Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /performance:
    get:
      summary: Get Performance
      operationId: get_performance_performance_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Performance'
                type: array
                title: Response Get Performance Performance Get
  /performance/{date}:
    get:
      summary: Get Performance For Date
      operationId: get_performance_for_date_performance__date__get
      parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
          title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Performance'
                title: Response Get Performance For Date Performance  Date  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transactions:
    get:
      summary: Get Transactions
      operationId: get_transactions_transactions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Transaction'
                type: array
                title: Response Get Transactions Transactions Get
  /transactions/{account_id}:
    get:
      summary: Get Transactions For Account
      operationId: get_transactions_for_account_transactions__account_id__get
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
                title: Response Get Transactions For Account Transactions  Account
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /metrics:
    get:
      summary: Get Metrics
      operationId: get_metrics_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metrics'
components:
  schemas:
    Account:
      properties:
        accountId:
          type: string
          pattern: ^PL-[a-f0-9]{32}$
          title: Account ID
        accountBalance:
          type: number
          title: Account Balance
        currentValue:
          type: number
          title: Current Value
        sharesOwned:
          type: number
          title: Shares Owned
        ownership:
          type: number
          title: Ownership
        returnOnInvestment:
          type: number
          title: Return On Investment
      type: object
      required:
      - accountId
      - accountBalance
      - currentValue
      - sharesOwned
      - ownership
      - returnOnInvestment
      title: Account
      description: "Represents an account.\n\nAttributes:\n    account_id: The ID\
        \ of the account associated with the transaction.\n    account_balance: The\
        \ balance of the account.\n    current_value: The current value of the account.\n\
        \    shares_owned: The number of shares owned in the account.\n    ownership:\
        \ The ownership percentage of the account.\n    return_on_investment: Return\
        \ on investment metrics."
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Performance:
      properties:
        date:
          type: string
          pattern: ^\d{4}-\d{2}-\d{2}$
          title: Date
        record:
          type: string
          pattern: ^\d+-\d+$
          title: Record
        returnOnInvestment:
          type: number
          title: Return On Investment
        shares:
          type: number
          title: Shares
        sharePrice:
          type: number
          title: Share Price
        totalBankroll:
          type: number
          title: Total Bankroll
        unitPrice:
          type: number
          title: Unit Price
        unitsWon:
          type: number
          title: Units Won
      type: object
      required:
      - date
      - record
      - returnOnInvestment
      - shares
      - sharePrice
      - totalBankroll
      - unitPrice
      - unitsWon
      title: Performance
      description: "Represents a performance record.\n\nAttributes:\n    date: The\
        \ date of the performance record.\n    record: A string representing the performance\
        \ record.\n    return_on_investment: The return on investment for the day.\n\
        \    shares: The number of shares held.\n    share_price: The price per share.\n\
        \    total_bankroll: The total bankroll for the day.\n    unit_price: The\
        \ price per unit.\n    units_won: The number of units won."
    Transaction:
      properties:
        date:
          type: string
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})?$
          title: Date
        accountId:
          type: string
          pattern: ^PL-[a-f0-9]{32}$
          title: Account ID
        amount:
          type: number
          title: Amount
        type:
          type: string
          pattern: ^(DEPOSIT|WITHDRAWAL)$
          title: Type
        shares:
          type: number
          title: Shares
        note:
          type: string
          pattern: ^(venmo|bank|check|cashapp|zelle|apple)$
          title: Note
      type: object
      required:
      - date
      - accountId
      - amount
      - type
      - shares
      - note
      title: Transaction
      description: "Represents a financial transaction.\n\nAttributes:\n    date:\
        \ The date of the transaction.\n    account_id: The ID of the account associated\
        \ with the transaction.\n    amount: The amount of the transaction.\n    type:\
        \ The type of the transaction (deposit, withdrawal).\n    shares: The number\
        \ of shares involved in the transaction.\n    note: Type of note (Venmo, Bank\
        \ Transfer, etc.)."
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Metrics:
      properties:
        currentSharePrice:
          type: number
          title: Current Share Price
        initialSharePrice:
          type: number
          title: Initial Share Price
        totalShares:
          type: number
          title: Total Shares
        totalBankroll:
          type: number
          title: Total Bankroll
      type: object
      required:
      - currentSharePrice
      - initialSharePrice
      - totalShares
      - totalBankroll
      title: Metrics
      description: "Represents metrics of the ledger.\n\nAttributes:\n    current_share_price:\
        \ The current price of a share (total_bankroll / total_shares).\n    initial_share_price:\
        \ The initial price of a share.\n    total_shares: The total number of shares\
        \ owned by all accounts.\n    total_bankroll: The total bankroll of the Ledger."